+++ posthog/temporal/ai/video_segment_clustering/schedule.py
@@ -1,53 +1,66 @@
-"""Schedule configuration for video segment clustering coordinator."""
+"""Schedule configuration for video segment clustering."""
 
 from django.conf import settings
 
 from temporalio.client import Client, Schedule, ScheduleActionStartWorkflow, ScheduleIntervalSpec, ScheduleSpec
 
 from posthog.temporal.ai.video_segment_clustering import constants
-from posthog.temporal.ai.video_segment_clustering.coordinator import VideoSegmentClusteringCoordinatorInputs
+from posthog.temporal.ai.video_segment_clustering.constants import get_proactive_tasks_team_ids
+from posthog.temporal.ai.video_segment_clustering.models import ClusteringWorkflowInputs
 from posthog.temporal.common.schedule import a_create_schedule, a_delete_schedule, a_schedule_exists, a_update_schedule
 
-SCHEDULE_ID = "video-segment-clustering-coordinator-schedule"
+
+def get_schedule_id(team_id: int) -> str:
+    """Get the schedule ID for a team."""
+    return f"video-segment-clustering-team-{team_id}"
 
 
-async def create_video_segment_clustering_schedule(client: Client):
-    """Create or update the schedule for video segment clustering.
+async def create_video_segment_clustering_schedules(client: Client):
+    """Create schedules for video segment clustering for each enabled team.
 
-    The coordinator discovers teams via feature flag and spawns child workflows
-    to cluster video segments for each team.
-
-    This schedule runs every 30 minutes.
+    Teams are enabled via the PROACTIVE_TASKS_TEAM_IDS environment variable.
+    Each team gets its own schedule running every 30 minutes.
     """
-    coordinator_schedule = Schedule(
-        action=ScheduleActionStartWorkflow(
-            "video-segment-clustering-coordinator",
-            VideoSegmentClusteringCoordinatorInputs(
-                lookback_hours=int(constants.DEFAULT_LOOKBACK_WINDOW.total_seconds() / 3600),
-                min_segments=constants.MIN_SEGMENTS_FOR_CLUSTERING,
-                max_concurrent_teams=constants.MAX_CONCURRENT_TEAMS,
-            ),
-            id=SCHEDULE_ID,
-            task_queue=settings.GENERAL_PURPOSE_TASK_QUEUE,
-        ),
-        spec=ScheduleSpec(intervals=[ScheduleIntervalSpec(every=constants.CLUSTERING_INTERVAL)]),
-    )
+    team_ids = get_proactive_tasks_team_ids()
 
-    if await a_schedule_exists(client, SCHEDULE_ID):
-        await a_update_schedule(client, SCHEDULE_ID, coordinator_schedule)
-    else:
-        await a_create_schedule(
-            client,
-            SCHEDULE_ID,
-            coordinator_schedule,
-            trigger_immediately=False,
+    for team_id in team_ids:
+        schedule_id = get_schedule_id(team_id)
+
+        team_schedule = Schedule(
+            action=ScheduleActionStartWorkflow(
+                "video-segment-clustering",
+                ClusteringWorkflowInputs(
+                    team_id=team_id,
+                    lookback_hours=int(constants.DEFAULT_LOOKBACK_WINDOW.total_seconds() / 3600),
+                    min_segments=constants.MIN_SEGMENTS_FOR_CLUSTERING,
+                ),
+                id=schedule_id,
+                task_queue=settings.GENERAL_PURPOSE_TASK_QUEUE,
+            ),
+            spec=ScheduleSpec(intervals=[ScheduleIntervalSpec(every=constants.CLUSTERING_INTERVAL)]),
         )
 
+        if await a_schedule_exists(client, schedule_id):
+            await a_update_schedule(client, schedule_id, team_schedule)
+        else:
+            await a_create_schedule(
+                client,
+                schedule_id,
+                team_schedule,
+                trigger_immediately=False,
+            )
 
-async def delete_video_segment_clustering_schedule(client: Client):
-    """Delete the video segment clustering coordinator schedule.
 
-    Args:
-        client: Temporal client
+async def delete_video_segment_clustering_schedules(client: Client):
+    """Delete all video segment clustering schedules.
+
+    Deletes schedules for all currently enabled teams.
     """
-    await a_delete_schedule(client, SCHEDULE_ID)
+    team_ids = get_proactive_tasks_team_ids()
+
+    for team_id in team_ids:
+        schedule_id = get_schedule_id(team_id)
+        try:
+            await a_delete_schedule(client, schedule_id)
+        except Exception:
+            pass  # Schedule may not exist
